+++ src/Ferrite.jl
@@ -35,6 +35,9 @@ abstract type BoundaryIndex end
 
 include("utils.jl")
 
+# Matrix/Vector utilities
+include("arrayutils.jl")
+
 # Interpolations
 include("interpolations.jl")
 
@@ -60,6 +63,7 @@ include(joinpath("Adaptivity", "AdaptiveCells.jl"))
 include("Dofs/DofHandler.jl")
 include("Dofs/MixedDofHandler.jl")
 include("Dofs/ConstraintHandler.jl")
+include("Dofs/DofRenumbering.jl")
 
 include("iterators.jl")