+++ test/runtests.jl
@@ -7,7 +7,11 @@ using Random
 using LinearAlgebra
 using SparseArrays
 
-include("test_octant.jl")
+const HAS_EXTENSIONS = isdefined(Base, :get_extension)
+if HAS_EXTENSIONS
+    import Metis
+end
+
 #include("test_utils.jl")
 #include("test_interpolations.jl")
 #include("test_cellvalues.jl")
@@ -24,4 +28,5 @@ include("test_octant.jl")
 ## include("test_notebooks.jl")
 #include("test_apply_rhs.jl")
 #include("test_examples.jl")
+include("test_p4est.jl")
 @test all(x -> isdefined(Ferrite, x), names(Ferrite))  # Test that all exported symbols are defined